+++ dev/cldr_backend.ex
@@ -1,5 +1,6 @@
 require Cldr.Backend
 require Cldr.Locale.Loader
+require MyApp.Gettext
 
 defmodule MyApp.Cldr do
   use Cldr,